kucoin-api/requirements.txt
python-kucoin
requests